Adds a mass-delete function
By clicking the checkbox button in the header bar, the user can check
which counters to delete, then click the delete button (replacing the
new button) to confirm deletion.

The search mode button is moved to the right side of the header bar to
make room for the new/delete/checkbox trio of buttons.