popout: Prevent keepalive bug
When a popout window is closed, it is normally just hidden. This
prevents the app from having stale references to destroyed windows, but
it also prevents the application from properly closing. Previously, the
app would try to close all popout windows associated with counters when
closing the main window, but that would not work for counters that had
been deleted after having their popup open, leaving the app in a zombie
state. This fixes that bug by destroying a counter's popout window when
deleting it.