Save counters more aggressively
Previously, counters would only save when exiting the app. This would
cause all changes to be lost if the user shut down their computer, or
suddenly lost power. This change makes it so that the current state of
the app will save itself about 1 second after the last change. Certain
changes, such as creating or deleting counters, cause immediate saves.
The app continues to save the state of counters when terminating the
app, but that should almost never be a factor.

Fixes #18

This commit also makes various tweaks to the code which don't affect the
app's function.