Bundles icons into the app
Tally was originally written with the assumption that the GNOME runtime
Flatpak would provide the system-provided icons shown in Icon Library,
but I later learned that this was not the case and that only a few icons
are shipped.

This commit adds a Gio.Resource containing all the necessary icons to
use the app, sourced from Icon Library. Because a few icons not in the
system-provided set were an improvement over the latter, the app also
switches to those (specifically, the icon for creating a standalone
window).

I may redraw these icons in the future, now that such a thing is
possible.

Fixes #19